Child command schedule on the same clock
If a command scheduled another command, due to the anonymous clock (used to freeze time).
The new command will be created on the anonymous clock and then fallback to default clock.

The solution is passing the origin clock down, called parentclock in the code, and when
scheduled new command, try to use it instead.